+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Add profile metadata @robgietema
 - Add @site endpoint @robgietema
 - Add upgrade profile functionality @robgietema
+- Add login by email @robgietema
 
 ### Bugfix